+++ izzet/assets/loaders/gltf/config.go
@@ -1,0 +1,11 @@
+package gltf
+
+const (
+	TextureCoordStyleOpenGL = 1
+)
+
+type TextureCoordStyle int
+
+type ParseConfig struct {
+	TextureCoordStyle TextureCoordStyle
+}
